+++ app/app-controllers/error-screen/errorController.js
@@ -41,6 +41,10 @@ define([
         });
     };
 
+    ErrorController.prototype.handleHardwareBackButtonPress = function() {
+        this.viewPlugin.events.trigger('back');
+    };
+
     ErrorController.prototype.show = function() {
         if (this.isShowing) {
             return;
@@ -90,11 +94,15 @@ define([
                 self.errorType = "noInternetConnection";
             }
 
+            // Only trigger the error modal if the active navigation view is
+            // the one that failed to load.
             if (isActiveItem()) {
                 self.viewPlugin.once('back', function() {
-                    self.hide();
-                    self._removeModalEvents();
-                    backHandler();
+                    if (self.canGoBack) {
+                        self.hide();
+                        self._removeModalEvents();
+                        backHandler();
+                    }
                 });
 
                 // Wait until the error page is loaded before showing
@@ -107,17 +115,9 @@ define([
                     var loadParams = {
                         errorContent: self.errorContent()
                     };
-
                     self.viewPlugin.trigger('error:should-load', loadParams);
                 });
             }
-
-            // Wait until the error page is loaded before showing
-            self.viewPlugin.on('error:loaded', function() {
-                self.show();
-            });
-
-            self.viewPlugin.navigate(ErrorConfig.url);
 
             // We allow all views that triggered this modal to listen for
             // `retry` so that they will reload when the error modal's retry