+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 ## To be released
+- Fix up error controller and error modal text
+- [Android] setup hardware back button and fix for both tab layout and left-nav layout
+- Add previewController to enable preview by shaking
+- Use global `build-js.sh` file from Astro
+
+## v0.16.1
 - Integrate the Mobify preview plugin
 - Welcome modal now relies on installationId to determine whether to show
 - Add Requirements to README, and "engines" field to package.json
+- Add Segmented Plugin
 - Replace WebView search bar with Search Bar Plugin
-- Use global `build-js.sh` file from Astro
+- Set iOS projects to 'iPhone' instead of 'Universal' by default
+- Upgrade to support Xcode 8 & Swift 2.3
+- Update build scripts to separate dependency build from app.js build
 
 ## v0.15.0
 - [iOS] Remove update bundle version build phase